Warn if method is shadowed and do not inherit mappings from parent classes
A check is made to see if the method has been defined explicitly
in the code before creating a trivial/const version.  For this
to work properly also needed to prevent parent class definitions
of _const_map and _trivial_map from appearing in the class
if they weren't explicitly nulled by the class.