Add python class object to properties definition and validate
Add an extra element to each entry defining a property.
The extra element is the actual Python type expected for the property.
This allows for the translated properties to be checked to ensure
they have the right type.

I considered trying to derive the class name from the current
class string (or using the type, derive the fully qualified
name) but it was easier to have the string form and the
python type explicitly at the risk of some duplication and
downstream breakage for people that define their own
properties.

The third element can be anything that can be passed to
isinstance().