Do a full install for travis
A developer install does not fully test the setup.py because
these checks were passing even when the setup.cfg was not
installing corrections files or the translators. The dev
install points back to the developer directory so is
ignoring the failure to install.  A full install will
install an egg and use the egg for subsequent
testing.