Remove PropertyList hack
Now that PropertySet supports __getitem__ and update there
is no need to convert them to OrderedDict before using the
headers.

For merging this now means that if you request to merge some PropertyList
headers you get back a PropertyList (which is critical if
you then want to pass the result to afw)

Still have to copy the input list since it is modified
in place.