Allow translator based header fix up to go wrong
It's possible that some headers are so broken that the
code trying to fix them up can't work out something.
A date is a good example since an error parsing a date
in a header will cause many date-based corrections to
fail.

Log the problem and hope that a file-based correction
is put in place that allows the translator code to
function properly.