allow extensions of the core set of properties
Some instruments have important properties beyond the core set
that have been defined, and we want to be able to extract those
as well. Added the concept of 'extensions' to both the
MetadataTranslator and ObservationInfo. This allows instruments
to define additional instrument-specific properties in their
appropriate MetadataTranslator, and these get picked up by the
ObservationInfo.

Unfortunately, extension properties have to get defined in the
ObservationInfo at runtime (because we don't know what they are
at compile time), which means we need to use a different scheme
for making the fields read-only, and there's no clear place to
hang their docstrings. But this is counteracted by the fact that
users don't need to choose a specific class to use, or define
the extensions themselves.