make subclassing of ObservationInfo easier
The ObservationInfo class gets post-processed to set up the
properties, which is not convenient for subclasses (which would
have to replicate the post-processing). Moved the post-processing
into a metaclass, so that this post-processing is done automatically
for ObservationInfo and all subclasses. This is more convenient
than the simpler class decorator solution, because the user doesn't
need to add the decorator.