Add ObservationInfo.observing_day_offset
This offset is what is used to calculate an observing day
from an observation date. There is also a class method allowing
the offset to be calculated from the observation date. Each
translator must declare an explicit value -- the default is
None and is equivalent to applying 0 offset when converting
a date time to an observing day.