Fix TELRA and TELDEC headers for Trifid DECam data
In attempting to process the data obtained for DM-45602:

http://astroarchive.noirlab.edu/
program number 2013A-0351
object trifid nebula

it was noted that the Telescope RA & Dec headers were offset by a
large enough distance that, when translated to the standard raw WCS,
none of the reference objects for calibration (astrometric and
photometric) overlaped the detector of interest.  This fix applies
the mean shift of the exposures of interest such that now they all
get much better raw WCS definitions and all calibration stages are
successful.